mpsl: fem: allow for seamless addition of new Front End Modules
To add support for a new Front End Module, implementation of already
supported Front End Modules must be modified. This commit improves that
integration process by allowing to add support for a new Front End
Module without any modifications of existing code.
